requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
docker==6.1.3
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
asyncpg==0.29.0
redis==5.0.1
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
python-multipart==0.0.6
jinja2==3.1.2
prometheus-client==0.19.0
structlog==23.2.0
croniter==2.0.1
apscheduler==3.10.4
httpx==0.25.2
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
psutil==5.9.6
watchdog==3.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
typer==0.9.0
